requirements.txt
# CREATE Table-Top Challenge 2026 — Requirements
# Python 3.12 | CUDA 12.6 (pre-installed)
# Install: pip install -r requirements.txt

monai==1.5.2
nibabel==5.3.3
SimpleITK==2.5.3
numpy==2.3.5
scipy==1.17.0
pandas==3.0.0
scikit-learn==1.8.0
PyYAML==6.0.3
matplotlib==3.10.8
tqdm==4.67.2
einops==0.8.0  # Required for SwinUNETR (rearrange operations)
tensorboard==2.16.2  # Required for Task 2 training logging
